refactor(table): replace expand-toggle and expandable-row
instead, use `<pf-tr expandable>`
