+++ elements/package.json
@@ -44,6 +44,13 @@
     "./pf-spinner/pf-spinner.js": "./pf-spinner/pf-spinner.js",
     "./pf-switch/BaseSwitch.js": "./pf-switch/BaseSwitch.js",
     "./pf-switch/pf-switch.js": "./pf-switch/pf-switch.js",
+    "./pf-table/pf-table.js": "./pf-table/pf-table.js",
+    "./pf-table/pf-thead.js": "./pf-table/pf-thead.js",
+    "./pf-table/pf-tbody.js": "./pf-table/pf-tbody.js",
+    "./pf-table/pf-tr.js": "./pf-table/pf-tr.js",
+    "./pf-table/pf-th.js": "./pf-table/pf-th.js",
+    "./pf-table/pf-td.js": "./pf-table/pf-td.js",
+    "./pf-table/pf-caption.js": "./pf-table/pf-caption.js",
     "./pf-tabs/BaseTab.js": "./pf-tabs/BaseTab.js",
     "./pf-tabs/BaseTabPanel.js": "./pf-tabs/BaseTabPanel.js",
     "./pf-tabs/BaseTabs.js": "./pf-tabs/BaseTabs.js",
@@ -56,7 +63,6 @@
     "./pf-tooltip/BaseTooltip.js": "./pf-tooltip/BaseTooltip.js",
     "./pf-tooltip/pf-tooltip.js": "./pf-tooltip/pf-tooltip.js",
     "./pf-popover/pf-popover.js": "./pf-popover/pf-popover.js",
-    "./pf-table/pf-table.js": "./pf-table/pf-table.js",
     "./react/*": "./react/*"
   },
   "publishConfig": {
