+++ elements/pf-table/demo/demo.css
@@ -1,8 +1,4 @@
 
-pf-expandable-row pf-table {
-  margin-bottom: -1px;
-}
-
 section {
   padding: 15px 30px;
 }
