--- elements/pf-table/pf-expand-toggle.css
@@ -1,15 +1,0 @@
-:host {
-    padding-left: 1.5rem;
-    position: absolute;
-    left: 0;
-    top: 1.5rem;
-}
-
-#icon {
-    transition: .2s ease-in 0s;
-    vertical-align: -0.125em;
-}
-  
-#button[aria-expanded="true"] #icon {
-    transform: rotate(90deg);
-}
