--- elements/pf-table/pf-expandable-row.css
@@ -1,9 +1,0 @@
-:host {
-    grid-auto-flow: row;
-    grid-template-columns: auto;
-    padding: 0 1.5rem;
-}
-
-:host(:not([expanded])) {
-    display: none;
-}
