+++ elements/pf-table/pf-expandable-rowset.css
@@ -1,22 +1,36 @@
-:host {
-    position: relative;
-}
-
-pf-expand-toggle {
-    z-index: 2;
-}
-
+:host,
 ::slotted(*) {
-    position: relative;
+  position: relative;
 }
 
 :host([expanded]) ::slotted(*)::before {
-    position: absolute;
-    top: -1px;
-    right: 0;
-    bottom: -1px;
-    left: 0;
-    content: "";
-    border-left: 3px solid #06c;
-    pointer-events: none;
+  position: absolute;
+  inset-block-start: -1px;
+  inset-inline-end: 0;
+  inset-block-end: -1px;
+  inset-inline-start: 0;
+  content: "";
+  border-inline-start: 3px solid #06c;
+  pointer-events: none;
+}
+
+::slotted(pf-table) {
+  margin-block-end: -1px;
+}
+
+#toggle-cell {
+  z-index: 2;
+  position: absolute;
+  inset-inline-start: 0;
+  inset-block-start: 1.5rem;
+  padding: 0;
+}
+
+#toggle-icon {
+  transition: .2s ease-in 0s;
+  vertical-align: -0.125em;
+}
+
+:host([expanded]) #toggle-icon {
+  transform: rotate(90deg);
 }
