+++ elements/pf-table/pf-expandable-rowset.ts
@@ -2,9 +2,17 @@ import { html } from 'lit';
 import { customElement } from 'lit/decorators/custom-element.js';
 import { property } from 'lit/decorators/property.js';
 
+import { ComposedEvent } from '@patternfly/pfe-core';
+
 import styles from './pf-expandable-rowset.css';
 
-import { ExpandChangeEvent, PfExpandableRow, PfTbody } from '@patternfly/elements/pf-table/pf-table.js';
+import { PfTbody, type PfTr } from '@patternfly/elements/pf-table/pf-table.js';
+
+export class ExpandChangeEvent extends ComposedEvent {
+  constructor(public expanded: boolean) {
+    super('change');
+  }
+}
 
 /**
  * Expandable rowset
@@ -23,7 +31,16 @@ export class PfExpandableRowset extends PfTbody {
 
   render() {
     return html`
-      <pf-expand-toggle></pf-expand-toggle>
+      <pf-td id="toggle-cell">
+        <pf-button id="toggle-button"
+                   aria-expanded=${this.expanded}
+                   plain
+                   @click=${this.#onClick}>
+          <pf-icon id="toggle-icon"
+                   icon="angle-right"
+                   size="md"></pf-icon>
+        </pf-button>
+      </pf-td>
       <slot></slot>
     `;
   }
@@ -32,12 +49,17 @@ export class PfExpandableRowset extends PfTbody {
     const { expanded } = event;
     this.expanded = expanded;
 
-    const expandableRow = Array.from(this.children)?.[1] as PfExpandableRow;
+    const expandableRow = this.querySelector<PfTr>('pf-tr[expandable]');
     if (expandableRow) {
       expandableRow.expanded = expanded;
     }
 
     event.stopPropagation();
+  }
+
+  #onClick() {
+    this.expanded = !this.expanded;
+    this.dispatchEvent(new ExpandChangeEvent(this.expanded));
   }
 }
